refine: chapter6,7,13
