+++ chapter13_computer_vision/13.3_bounding-box.py
