+++ chapter13_computer_vision/13.4_anchor.py
