+++ chapter13_computer_vision/13.6_SSD.py
