+++ chapter13_computer_vision/13.7_semantic-segmentation.py
