+++ chapter13_computer_vision/13.8_transposed_conv.py
