--- chapter7/7.3_NiN.py
@@ -1,27 +1,0 @@
-import torch
-from torch import nn
-from d2l import torch as d2l
-
-def nin_block(in_channels, out_channels, kernel_size, stride, padding): # kernel_size, strides, padding给第一个卷积层
-    return nn.Sequential(
-        nn.Conv2d(in_channels, out_channels, kernel_size, stride, padding), nn.ReLU(),
-        nn.Conv2d(out_channels, out_channels, kernel_size=1), nn.ReLU(),
-        nn.Conv2d(out_channels, out_channels, kernel_size=1), nn.ReLU()
-    )
-
-net = nn.Sequential(
-    nin_block(1, 96, kernel_size=11, stride=4, padding=0),
-    nn.MaxPool2d(kernel_size=3, stride=2),
-    nin_block(96, 256, kernel_size=5, stride=1, padding=2),
-    nn.MaxPool2d(kernel_size=3, stride=2),
-    nin_block(256, 384, kernel_size=3, stride=1, padding=1),
-    nn.MaxPool2d(kernel_size=3, stride=2),
-    nin_block(384, 10, kernel_size=3, stride=1, padding=1),
-    nn.AdaptiveAvgPool2d((1, 1)), # 自适应输入尺寸，把每个channel都调整为(1, 1)的输出（也就是输出每个channel的平均值）
-    nn.Flatten()
-)
-
-X = torch.randn(size=(1, 1, 224, 224))
-for layer in net:
-    X = layer(X)
-    print(f'{layer.__class__.__name__} output shape:\t{X.shape}')
