--- chapter7/7.5_batch_norm.py
@@ -1,13 +1,0 @@
-import torch
-from torch import nn
-from d2l import torch as d2l
-
-def batch_norm(X, gamma, beta, moving_mean, moving_var, eps, moment):
-    if not torch.is_grad_enabled():
-        X_hat = (X - moving_mean) / torch.sqrt(moving_var + eps)
-    else:
-        assert len(X.shape) in (2, 4)   # 2 --> Linear   4 --> Conv2d
-        if len(X.shape) == 2:
-            mean = X.mean(dim=0)
-            var = ((X - mean)**2).mean(dim=0)
-        else:
