+++ chapter7_conv_modern/7.6_ResNet.py
@@ -1,0 +1,45 @@
+import torch
+from torch import nn
+from torch.nn import functional as F
+from d2l import torch as d2l
+
+class Residual(nn.Module):
+    def __init__(self, input_channels, num_channels, use_1x1conv=False, strides=1):
+        super().__init__()
+        self.conv1 = nn.Conv2d(input_channels, num_channels, kernel_size=3, padding=1, stride=strides)
+        self.conv2 = nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1)
+
+        if use_1x1conv: # 表示有无残差块，如果有残差就要把输入的channel变换到输出的channel，这样才能相加
+            self.conv3 = nn.Conv2d(input_channels, num_channels, kernel_size=1, stride=strides)
+        else:
+            self.conv3 = None
+
+    def forward(self, X):
+        Y = F.relu(self.bn1(self.conv1(X)))
+        Y = self.bn2(self.conv2(Y))
+        if self.conv3:  # 处理残差
+            X = self.conv3(X)
+        Y += X
+        return F.relu(Y)
+
+
+b1 = nn.Sequential(nn.Conv2d(1, 64, kernel_size=7, stride=2, padding=1),
+                   nn.BatchNorm2d(64), nn.ReLU(),
+                   nn.MaxPool2d(kernel_size=3, stride=2, padding=1))
+
+def resnet_block(input_channels, num_channels, num_residuals, first_block=False):
+    blk= []
+    for i in range(num_residuals):
+        if i == 0 and not first_block:
+            blk.append(
+                Residual(input_channels, num_channels, use_1x1conv=first_block, strides=2)
+            )
+        else:
+            blk.append(Residual(num_channels, num_channels))
+    return blk
+
+
+b2 = nn.Sequential(*resnet_block(64, 64, 2, first_block=True))
+b3 = nn.Sequential(*resnet_block(64, 128, 2))
+b4 = nn.Sequential(*resnet_block(128, 256, 2))
+b5 = nn.Sequential(*resnet_block(256, 512, 2))
